% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_table.R
\name{characteristic_table_generator}
\alias{characteristic_table_generator}
\title{Characteristic Table Generator}
\usage{
characteristic_table_generator(
  data,
  by = NULL,
  boolean_list = NULL,
  sd_list = NULL,
  IQR_list = NULL,
  order = NULL,
  normal_policy = "wilcox"
)
}
\arguments{
\item{data}{data.frame, the data.frame working on}

\item{by}{character, the column name used for stratifying}

\item{boolean_list}{optional character vector, the column name which is treated as boolean variable}

\item{sd_list}{optional character vector, the column name which is treated as normal distributed continuous variable}

\item{IQR_list}{optional character vector, the column name which is treated as non-normal distributed continuous variable}

\item{order}{optional character vector, if specified, the rows will follow the order provided}

\item{normal_policy}{str, for variables in sd_list, if "wilcox", run stats::wilcox.test, if "t-test", run stats::t.test}
}
\value{
data.frame, the cleaned result
}
\description{
Wrapper function cleaning the format of generated statistic
}
