% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_table.R
\name{characteristic_wrapper}
\alias{characteristic_wrapper}
\title{Single column inspector}
\usage{
characteristic_wrapper(data, variable_list, by, mode = "boolean")
}
\arguments{
\item{data}{data.frame, the data.frame working on}

\item{variable_list}{character vector, the vector of variables included}

\item{by}{character, the column name used for stratifying}

\item{mode}{character, the mode generating result, can be "boolean", "continuous_sd', "continuous_IQR"}
}
\value{
data.frame, the cleaned result
}
\description{
a wrapper function cleaning the format from generated statistic
}
