% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{clean_correlation_matrix}
\alias{clean_correlation_matrix}
\title{Function cleaning input and computing correlation matrix}
\usage{
clean_correlation_matrix(data, variable = NULL)
}
\arguments{
\item{data}{data.frame, the data table.}

\item{variable}{optional character vector, the list of variable included in computation, all columns by default}
}
\value{
a list with elements r, the matrix of correlations, n the matrix of number of observations used in analyzing each pair of variables, P, the asymptotic P-values, and type.
}
\description{
If no variable list provided, the function will clean all the value as numerical matrix
and compute the correlation matrix
}
