% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nih_equation.R
\name{compute_nih_ldl}
\alias{compute_nih_ldl}
\title{NIH 2 equation, also known as the Sampson-NIH equation}
\usage{
compute_nih_ldl(tchol, hdl, trig)
}
\arguments{
\item{tchol, }{float, total cholesterol}

\item{hdl, }{float HDL level}

\item{trig, }{float Triglyceride level}
}
\value{
float, the LDL level
}
\description{
Compute the LDL(calc) based on NIH equation 2
}
