% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{correlation_matrix_dataframe}
\alias{correlation_matrix_dataframe}
\title{Function wrapper compute pairwise correlation matrix and return as a data.frame}
\usage{
correlation_matrix_dataframe(data, variable = NULL, p_threshold = 1)
}
\arguments{
\item{data}{data.frame, the data table.}

\item{variable}{optional character vector, the list of variable included in computation, all columns by default}

\item{p_threshold}{optional float, the only the rows with p-value less than this value will be included, default 1 (all included)}
}
\value{
data.frame, a data.frame with two columns: corr and p_value
}
\description{
If no variable list provided, the function will clean all the value as numerical matrix
and compute the correlation matrix
}
