% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{correlation_matrix_plot}
\alias{correlation_matrix_plot}
\title{Function wrapper compute pairwise correlation matrix and return as a visualization format}
\usage{
correlation_matrix_plot(
  data,
  variable = NULL,
  p_threshold = NULL,
  font_size = 5
)
}
\arguments{
\item{data}{data.frame, the data table.}

\item{variable}{optional character vector, the list of variable included in computation, all columns by default}

\item{p_threshold}{optional float, the only the rows with p-value less than this value will be included, Null default(all included)}

\item{font_size}{optional int, the size of row/column labels, default 5}
}
\value{
data.frame, a data.frame with two columns: corr and p_value
}
\description{
If no variable list provided, the function will clean all the value as numerical matrix
and compute the correlation matrix
}
