% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{create_multi_color_palette}
\alias{create_multi_color_palette}
\title{Create a non-white color palette for ggplot2 based on Polychrome::glasbey.colors}
\usage{
create_multi_color_palette(num_color)
}
\arguments{
\item{num_color, }{int, the number of color need}
}
\value{
list of character, the color list
}
\description{
Generate a color palette for ggplot2 with specified number of color, white is intentionally removed.
}
