% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distinguishable_palette}
\alias{distinguishable_palette}
\title{Wrapper creating a distinguishable, non-white color palette for ggplot2 based on Polychrome::glasbey.colors}
\usage{
distinguishable_palette(num_color, pre_specified_palette = NULL)
}
\arguments{
\item{num_color, }{int, the number of color need}

\item{pre_specified_palette, }{vector of str, the palette provided, if provided, will directly output this vector}
}
\value{
list of str, the color list
}
\description{
Generate a color palette for ggplot2 with specified number of color, white is intentionally removed, added codes dealt with 1 or 2 color(s).
}
