% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_rank.R
\name{enrichment_rank}
\alias{enrichment_rank}
\title{Distribution illustration for proteomics/lipidomics, etc}
\usage{
enrichment_rank(
  data,
  formal_name,
  id = NULL,
  save_name_col = NULL,
  formal_name_col = NULL,
  sort_policy = "median",
  scale = "raw"
)
}
\arguments{
\item{data}{data.frame carries the enrichment}

\item{formal_name}{data.frame, the table including the formal name and save name of the output}

\item{id}{optional str, the column name in data referring to the subject id, by default the first column of data}

\item{save_name_col}{optional str, the column name in formal_name, taking the save_name from pipeline output}

\item{formal_name_col}{optional str, the column name in formal_name, taking the formal name in illustration}

\item{sort_policy, }{optional str, the sort policy of enrichments, by default "median", "mean" provided}

\item{scale, }{optional str, the scale of enrichments, by default "raw", "logged" provided as log10(x+1)}
}
\value{
a list with two element, "visualization" is a ggplot2 instance as a scatter plot, "data" is the source data
}
\description{
A wrapper cleaning the output from
}
