% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{flatten_correlation_matrix}
\alias{flatten_correlation_matrix}
\title{Function cleaning correlation matrix to pairwise dataframe}
\usage{
flatten_correlation_matrix(corr_matrix, p_matrix)
}
\arguments{
\item{corr_matrix}{matrix, a numerical matrix, the correlation matrix}

\item{p_matrix}{matrix, a numerical matrix, the p-value matrix}
}
\value{
data.frame, a data.frame with two columns: corr and p_value
}
\description{
This function takes a correlation matrix and a p-value matrix from correlation matrix computing
clean the matrix to pairwise terms
credit to https://www.sthda.com/english/wiki/correlation-matrix-a-quick-start-guide-to-analyze-format-and-visualize-a-correlation-matrix-using-r-software
}
