% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_table.R
\name{generate_boolean_ratio_output}
\alias{generate_boolean_ratio_output}
\title{Characteristics for boolean variable}
\usage{
generate_boolean_ratio_output(variable, df, by)
}
\arguments{
\item{variable}{The name of column in table df}

\item{df}{the table working on}

\item{by}{str, stratification variable name}
}
\value{
data.frame, frequency and relative frequency for the second term in table() and prop.table which is designed to be 1 (positive indicator for a disease against as negative)
}
\description{
A function extract boolean 1/total(percentage) format from a column named variable from table df, when by has and only has 2 classes, run a chi-square test
}
