% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_table.R
\name{generate_continuous_IQR_output}
\alias{generate_continuous_IQR_output}
\title{Characteristics for non-normality distributed variable}
\usage{
generate_continuous_IQR_output(variable, df, by)
}
\arguments{
\item{variable}{str, The name of column in table df}

\item{df}{data.frame, the table working on}

\item{by}{str, stratification variable name}
}
\value{
data.frame, the characteristic data median(IQR), when by has and only has 2 classes, run a Mann Whitney U test
}
\description{
Extract median(IQR) format from a column named variable from table df,
when "by" has and only has 2 classes, run a Mann-Whitney U test
}
