% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_odd_ratio}
\alias{generate_odd_ratio}
\title{Compute OR from effect size in logistic regression}
\usage{
generate_odd_ratio(
  data,
  effect_size_name = "Effect Size",
  odd_ratio_name = "Odd Ratio"
)
}
\arguments{
\item{data}{data.frame, the table work on}

\item{effect_size_name}{str, the name of column of effect size}

\item{odd_ratio_name}{str, the name of new column generated as odd ratio}
}
\value{
data.frame with an additional odds_ratio
}
\description{
Generate an "Odd Ratio" columns from effect size, and put it after effect size
}
