% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_and_normality.R
\name{inspect_distribution}
\alias{inspect_distribution}
\title{Wrapper quickly check the distribution from Shapiro-Wilk test and visualization}
\usage{
inspect_distribution(data, variable, by = NULL, palette_color = NULL)
}
\arguments{
\item{data, }{data.frame, the data table with original data}

\item{variable, }{str or vector of str, the column name of variable of interest}

\item{by, }{optional str, if given, the column by which data will be stratified}

\item{palette_color, }{optional vector of str, list of colors used}
}
\value{
a list with two element, "visualization" is a list of ggplot instance with the histogram and density plot, "shapiro-wilk_test" is a data.frame
}
\description{
Wrapper quickly check the distribution from Shapiro-Wilk test and visualization
}
