% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associations.R
\name{linear_regression}
\alias{linear_regression}
\title{Ordinary Linear Regression}
\usage{
linear_regression(
  data,
  variable,
  response,
  adjustments,
  mode = "table",
  with_power = FALSE,
  variable_distribution = "normal"
)
}
\arguments{
\item{data}{data.frame the table working on}

\item{variable}{str, the name of column of interested variable in table data}

\item{response}{str, the name of column of target variable in table data}

\item{adjustments}{vector of str, the list of adjustment variables in the table data}

\item{mode}{str, when "summary", only output the coefficient corresponding to column specified in variable}

\item{with_power}{boolean, when set to TRUE, will call power.t.regression function computing the power and sample size at 80 percent power, default FALSE to save computation burden}

\item{variable_distribution}{str, either "normal" or "binary", the distribution of variable of interest specified, only necessary when with_power = TRUE, default "normal"}
}
\value{
data.frame, if mode summary, only the beta and p-value for variable specified in "variable", otherwise the general summary()
}
\description{
A wrapper creating linear regression formula and clean the output
}
