% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Wrapper for a Manhattan plot}
\usage{
manhattan_plot(
  data,
  pos,
  phenotype,
  p_value,
  p_value_threshold = 0.05,
  formal_name = NULL,
  coding = NULL,
  phenotype_list = NULL,
  highlight_region = NULL,
  palette_color = NULL,
  coord_ratio = 2500/1
)
}
\arguments{
\item{data, }{data.frame, The table working on}

\item{pos, }{string, the name of column of position of the variant}

\item{phenotype, }{str, the name of column of phenotype the variant associated to}

\item{p_value, }{str, the name of column of p-values}

\item{p_value_threshold, }{float, the threshold for significance, default 0.05}

\item{formal_name, }{optional str, the name of column of formal name of each variant, if not specified, will use pos}

\item{coding, }{optional str, (currently not used), the name of column of exonic vs intronic,}

\item{phenotype_list, }{optional vector of str, the list of phenotypes which is put into consideration}

\item{highlight_region, }{optional list of vector of str, in list(c(start, end)...) format, the list of highlighted region}

\item{palette_color, }{list or str, pre-specfied colors used for different phenotypes}

\item{coord_ratio, }{float, the coordinate ratio for the figure}
}
\value{
ggplot2 instance, the Manhattan plot
}
\description{
A wrapper function cleans column format for a Manhattan plot
}
