% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_table.R
\name{multi_line_summary}
\alias{multi_line_summary}
\title{Multi-column Summary}
\usage{
multi_line_summary(data)
}
\arguments{
\item{data}{data.frame, the data.frame working on}
}
\value{
data.frame, the cleaned result
}
\description{
Wrapper function output multiple column base::summary() result into data.frame
I really cannot understand why R returns that in that weird way.
}
