% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directed_acyclic_graph.R
\name{multi_omics_directed_acyclic_graph}
\alias{multi_omics_directed_acyclic_graph}
\title{Generate and preview a directed acyclic graph for multi-omics result}
\usage{
multi_omics_directed_acyclic_graph(
  data,
  from,
  to,
  beta,
  p_value,
  from_class,
  to_class,
  class_level = NULL,
  color_palette = NULL
)
}
\arguments{
\item{data}{data.frame, the data input}

\item{from}{str, the column name for the starting point}

\item{to}{str, the column name for the end point}

\item{beta}{str, the column name for the beta}

\item{p_value, }{str, the column name for p-value in the association study}

\item{from_class, }{str, the column name for the class label of from column if specified}

\item{to_class, }{str, the column name for the class label of to column if specified}

\item{class_level, }{optional vector of str, if need to specify a order for classes defined in from_class and to_class}

\item{color_palette, }{optional vector of str, the color for each class, the package will prepare a default color if not specified.}
}
\value{
A list of result with following attributes:

 * nodes data.frame, the nodes table
 * edges data.frame, the edges table
 * illustration_network visNetwork instance, the network in illustration
 * raw_network visNetwork instance, the network without any visNetwork::visPhysics settings
}
\description{
A wrapper prepare the data for a directed acyclic graph for multi-omics data, and preview it
}
