% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_and_normality.R
\name{normality_table}
\alias{normality_table}
\title{Shapiro-Wilk test on each variable}
\usage{
normality_table(data, by = NULL)
}
\arguments{
\item{data, }{data.frame, the data table to be tested}

\item{by, }{optional str, the column by which data will be stratified}
}
\value{
data.frame, a data frame including all the results
}
\description{
For a data.frame, run Shapiro-Wilk test on each column, if a "by" is specified, it will run a stratified Shapiro-wilk test by this column
}
