% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associations.R
\name{OLS_wrapper_}
\alias{OLS_wrapper_}
\title{Ordinary Linear Regression with multi variable of interest}
\usage{
OLS_wrapper_(
  data,
  response,
  adjustments,
  variable_of_interest,
  variable_of_interest_formal_name,
  with_power = TRUE,
  variable_distribution = NULL
)
}
\arguments{
\item{data}{data.frame the table working on}

\item{response}{str, the name of column of target variable in table data}

\item{adjustments}{vector of str, the list of adjustment variables in the table data}

\item{variable_of_interest}{vector of str, the names of columns of interested variable in table data}

\item{variable_of_interest_formal_name}{vector of str, the output names of interested variable in result}

\item{with_power}{boolean, when set to TRUE, will call power.t.regression function computing the power and sample size at 80 percent power, default FALSE to save computation burden}

\item{variable_distribution}{vector of str, either "normal" or "binary", the distribution of variable of interest specified, only necessary when with_power = TRUE, default "normal", should be 1-1 corresponding to variable of interest}
}
\value{
data.frame, one variable in variable_of_interest as a row, with it's beta and p-value
}
\description{
A wrapper running multiple linear_regression() for a list of variables of interest
}
