% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_pairwise_comparison.R
\name{pairwise_boolean_table}
\alias{pairwise_boolean_table}
\title{Pairwise distribution comparison report for boolean(0/1) variables}
\usage{
pairwise_boolean_table(data, variable, by, adjustment_method = "fdr")
}
\arguments{
\item{data}{data.frame, the source data}

\item{variable}{str, the variable interested in}

\item{by}{str, the categorical variable stratifying the dataset}

\item{adjustment_method}{optional str, the correction adjusting pairwise, Benjamini-Hochberg FDR by default}
}
\value{
data.frame, the pairwise result for variable(s) stratified by "by" in "data"
}
\description{
A wrapper running allowing multiple variable of interest, will run the result in an interative manner
}
