% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_pairwise_comparison.R
\name{pairwise_t_test_report}
\alias{pairwise_t_test_report}
\title{A wrapper preparing pairwise distribution comparison report normally distributed variables}
\usage{
pairwise_t_test_report(
  data,
  variable,
  by,
  p_threshold = 0.05,
  adjustment_method = "fdr"
)
}
\arguments{
\item{data}{data.frame, the source data}

\item{variable}{str or vector of str, the variable you are interested in}

\item{by}{str, the categorical variable stratifying the dataset}

\item{p_threshold}{optional float, the significance threshold, 0.05 by default}

\item{adjustment_method}{optional str, the correction adjusting pairwise, Benjamini-Hochberg FDR by default}
}
\value{
a list with following attributes:
* result: data.frame complete pairwise result.
* significant_result: data.frame which only include the significant result
* plot: ggplot2 instance, the figure illustrating the significant result.
}
\description{
This wrapper will run a stats::pairwise.t.test for each element mentioned in "variable" while stratified across different value in "by" column and prepare a barchart for illustration.
}
