% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms_illustration.R
\name{rms_illustration_3d}
\alias{rms_illustration_3d}
\title{3d visualization for rms fit object}
\usage{
rms_illustration_3d(
  fit,
  datadist,
  variable,
  target_name = "target",
  density = 50
)
}
\arguments{
\item{fit, }{rms model fit result}

\item{datadist, }{rms datadist() result}

\item{variable, }{vector of str with length 2, the two variables on x and y axis}

\item{target_name, }{str, the axis label for prediction on z axis, default "target"}

\item{density, }{int or vector of int with length 2, the number of points on x and y axis, if only given one number will deployed on both axes, default 50}
}
\value{
plotly::plot_ly instance, the 3d output for the model
}
\description{
A wrapper generating plotly figures for rms result in 3d surface, mimicking the behaviour for ggrmsMD in 2D
}
