% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_bar_chart.R
\name{trend_bar_chart}
\alias{trend_bar_chart}
\title{A fill bar chart for composition over a specific variable}
\usage{
trend_bar_chart(
  data,
  id,
  by,
  variable = NULL,
  legend_title = "Component",
  palette_color = NULL
)
}
\arguments{
\item{data, }{data.frame, The table working on}

\item{id, }{str, the name of column of subject id}

\item{by, }{str, the name of column of variable on x}

\item{variable, }{optional vector of str, the name of column of composition components if specified, otherwise will use all the other variables}

\item{legend_title, }{str, the legend title in the figure}

\item{palette_color, }{list or str, pre-specfied colors used for different phenotypes}
}
\value{
ggplot2 instance, the plot output
}
\description{
This function returns a filled bar chart illustrate the trend of a composition over the change of a specific variable
}
