% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_table.R
\name{characteristic_wrapper}
\alias{characteristic_wrapper}
\title{Single column inspector}
\usage{
characteristic_wrapper(
  data,
  variable_list,
  by,
  mode = "boolean",
  normal_policy = "wilcox"
)
}
\arguments{
\item{data}{data.frame, the data.frame working on}

\item{variable_list}{character vector, the vector of variables included}

\item{by}{character, the column name used for stratifying}

\item{mode}{character, the mode generating result, can be "boolean", "continuous_sd", "continuous_IQR"}

\item{normal_policy}{str, for "continuous_sd" mode, if "wilcox", run stats::wilcox.test, if "t-test", run stats::t.test}
}
\value{
data.frame, the cleaned result
}
\description{
Wrapper function cleaning the format of generated statistic
}
