% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_table.R
\name{generate_continuous_sd_output}
\alias{generate_continuous_sd_output}
\title{Characteristics for normality distributed variable}
\usage{
generate_continuous_sd_output(variable, df, by)
}
\arguments{
\item{variable}{The name of column in table df}

\item{df}{the table working on}
}
\value{
data.frame, the characteristic data mean(sd), when by has and only has 2 classes, run a Mann Whitney U test
}
\description{
A function extract mean(sd) format from a column named variable from table df, when by has and only has 2 classes, run a Mann Whitney U test (I didn't deploy 2 sample t-test for I cannot assume sample size and variance)
}
