% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associations.R
\name{logistic_regression}
\alias{logistic_regression}
\title{Logistics Linear Regression}
\usage{
logistic_regression(data, variable, response, adjustments, mode = "table")
}
\arguments{
\item{data}{data.frame the table working on}

\item{variable}{str, the name of column of interested variable in table data}

\item{response}{str, the name of column of target variable in table data}

\item{adjustments}{vector of str, the list of adjustment variables in the table data}

\item{mode}{str, when "summary", only output the coefficient corresponding to column specified in variable}
}
\value{
data.frame, if mode summary, only the beta and p-value(z-value) for variable specified in "variable", otherwise the general summary()
}
\description{
A wrapper creating logistic regression formula and clean the output
}
