% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associations.R
\name{logistic_wrapper}
\alias{logistic_wrapper}
\title{Logistics Linear Regression wrapper with multi variable of interest and stratification}
\usage{
logistic_wrapper(
  data,
  response,
  adjustments,
  by = NULL,
  variable_of_interest = NULL,
  variable_of_interest_formal_name = NULL
)
}
\arguments{
\item{data}{data.frame the table working on}

\item{response}{str, the name of column of target variable in table data}

\item{adjustments}{vector of str, the list of adjustment variables in the table data}

\item{by}{str, the name of column of stratified variable in table data}

\item{variable_of_interest}{vector of str, the names of columns of interested variable in table data}

\item{variable_of_interest_formal_name}{vector of str, the output names of interested variable in result}
}
\value{
data.frame, one variable in variable_of_interest as a row, with it's beta and p-value
}
\description{
A wrapper running multiple linear_regression() for a list of variables of interest, stratification function provided
}
