% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associations.R
\name{logistic_wrapper_}
\alias{logistic_wrapper_}
\title{Logistics Linear Regression with multi-variable of interest}
\usage{
logistic_wrapper_(
  data,
  response,
  adjustments,
  variable_of_interest,
  variable_of_interest_formal_name
)
}
\arguments{
\item{data}{data.frame the table working on}

\item{response}{str, the name of column of target variable in table data}

\item{adjustments}{vector of str, the list of adjustment variables in the table data}

\item{variable_of_interest}{vector of str, the names of columns of interested variable in table data}

\item{variable_of_interest_formal_name}{vector of str, the output names of interested variable in result}
}
\value{
data.frame, one variable in variable_of_interest as a row, with it's beta and p-value
}
\description{
A wrapper running multiple logistics_regression() for a list of variables of interest
}
