% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_data_clean}
\alias{manhattan_data_clean}
\title{Clean format for a Manhattan plot}
\usage{
manhattan_data_clean(
  data,
  pos,
  phenotype,
  p_value,
  formal_name = NULL,
  coding = NULL,
  phenotype_list = NULL
)
}
\arguments{
\item{data, }{data.frame, The table working on}

\item{pos, }{string, the name of column of position of the variant}

\item{phenotype, }{str, the name of column of phenotype the variant associated to}

\item{p_value, }{str, the name of column of p-values}

\item{formal_name, }{str, the name of column of formal name of each variant}

\item{coding, }{str, the name of column of exonic vs intronic, currently not used}

\item{phenotype_list, }{vector of str, the list of phenotypes which is put into consideration}
}
\value{
data.frame, the table cleaned
}
\description{
A function cleans column format for a Manhattan plot, in which "phenotype" and "coding" should be factor, "pos" and "p-value" should be numerical, formal name should be character
}
