% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrorstoolkit-package.R
\docType{package}
\name{mirrorstoolkit-package}
\alias{mirrorstoolkit}
\alias{mirrorstoolkit-package}
\title{mirrorstoolkit: Toolkit for biomedical paper output preparation}
\description{
The code accumulated in the practice for most often used functionalities in biomedical studies.
}
\author{
\strong{Maintainer}: Yihao Li \email{alexliyihao@gmail.com} (\href{https://orcid.org/0009-0009-8722-3437}{ORCID})

Other contributors:
\itemize{
  \item Gissette Reyes-Soffer \email{gr2104@cumc.columbia.edu} (\href{https://orcid.org/0000-0002-6746-4569}{ORCID}) [funder]
  \item Badri Vardarajan \email{bnv2103@cumc.columbia.edu} (\href{https://orcid.org/0000-0002-5560-4085}{ORCID}) [funder]
}

}
\keyword{internal}
